pyproject.toml
[tool.poetry]
name = "MythicPlusBot"
version = "0.5.0"
description = "A discord bot for tracking mythic+ runs."
authors = ["Eric Tossell <eric@tossell.ca>"]
# readme = "README.md"
# license = "BSD"
packages = [
  { include = "bot" }
]

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.4.1"
tqdm = "^4.66.1"
ratelimit = "^2.2.1"
httpx = "^0.26.0"
motor = "^3.3.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
matplotlib = "3.7.1"
dnspython = "^2.4.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
pandas = "2.0.2"
adjusttext = "^1.0.4"
datetime = "^5.4"
numpy = "1.24.2"

[tool.poetry.scripts]
mythicPlusBot = "bot.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

